alignment fast paths for vectorized loads and many refactoring for simplification. v0.1.2

+++ perfs/cuda/julia/scan/bench_cub_kokkos.jl
@@ -23,4 +23,4 @@ function run_kokkos_scan_benchmark(exe::String;
 end
 
 kokkos_exe::String = "$(@__DIR__)/../../cuda_cpp/kokkos/build/bench_kokkos_scan"
-run_kokkos_scan_benchmark(kokkos_exe)
+run_kokkos_scan_benchmark(kokkos_exe, dtype=Float64)
